DB migration: Fix error on first run
When first importing the table migration stored procedure, we'd get:

    _mysql_exceptions.ProgrammingError) (2014, "Commands out of
    sync; you can't run this command now")

... for the same reason as in commit 8ed66a7431: there are two queries
which produce output within the imported file, so the driver cannot
properly process both of them.

The fix is to simplify the SQL file to 1 query: the stored procedure
creation query (on the OpenSIPS side), and catch the "SP already exists"
exception (on the opensips-cli side).