database: Add auto-user creation logic for PostgreSQL
Identical usage as in 70232defb.

This commit also removes the unnecessary "role management" commands
specific to Postgres.  opensips-cli is not meant to be a Postgres role
manager, and any roles (or backend-specific permission abstractions) it
needs should be transparently managed.  Users that needs to
add/get/edit/delete roles may happily use the Postgres CLI / queries
for this.

Related to #19