database:  Require Zero (0) Configuration.
Similarly to how opensipsdbctl used to work, the database module now
simply works out of the box for MySQL/PostgreSQL, with no configuration
required.

During DB creation, the user is now prompted to type in the admin
account password, if one is not already provided via the CLI
configuration file.